requirements.txt
# Crawler
Scrapy==1.6.0

# sql
mysqlclient==1.3.12

# config parser
configparser==3.7.1

# splash
scrapy-splash==0.7.2

# python .env
python-dotenv==0.1.0

# parser date
dateparser==0.7.1


